@hubgeter 2024-07-10
run buildall