@hubgeter 2024-07-11
run buildall